services/api/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
httpx
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
